Automatic "Data Recovery"
The utility targets the hardware of the running Linux machine and attempts to find a connected Windows/MacOSX hard drive attached via disk reader and produce a backup of the personal files contained on the drive.  I wrote this for one of my former employers and it is still being used today.  This does not actually recover data that has been deleted (though I have solved this using binary grep on the device block), but rather simplifies the process of recovering data for customers in need.